backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.27.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
marshmallow==3.26.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.4
sqlparse==0.5.3
transbank-sdk==6.1.0
tzdata==2025.2
urllib3==2.5.0
